fix: BrowserTestRunResult
