+++ packages/core/src/core/browserLoader.ts
@@ -1,36 +1,13 @@
 import { createRequire } from 'node:module';
 import { pathToFileURL } from 'node:url';
-import type { ListCommandResult, TestFileResult, TestResult } from '../types';
+import type {
+  BrowserTestRunOptions,
+  BrowserTestRunResult,
+  ListCommandResult,
+} from '../types';
 import { color, logger } from '../utils';
 
-/**
- * Options for running browser tests.
- */
-export interface BrowserTestRunOptions {
-  /**
-   * If true, browser mode will not call onTestRunEnd reporter hook.
-   * This allows the caller to unify reporter output with node mode tests.
-   */
-  skipOnTestRunEnd?: boolean;
-}
-
-/**
- * Result from running browser tests.
- */
-export interface BrowserTestRunResult {
-  /** Test file results */
-  results: TestFileResult[];
-  /** Individual test case results */
-  testResults: TestResult[];
-  /** Duration information */
-  duration: {
-    totalTime: number;
-    buildTime: number;
-    testTime: number;
-  };
-  /** Whether the test run had failures */
-  hasFailure: boolean;
-}
+export type { BrowserTestRunOptions, BrowserTestRunResult } from '../types';
 
 /**
  * Type definition for the @rstest/browser package exports.
