+++ packages/core/src/types/index.ts
@@ -1,4 +1,5 @@
 export type * from './api';
+export type * from './browser';
 export type * from './config';
 export type * from './core';
 export type * from './coverage';
